requirements.txt
langchain-core>=0.1.31
langchain-community>=0.0.28
langchain-chroma>=0.0.4
python-dotenv==1.0.1
chromadb>=0.4.15
streamlit==1.32.2
pydantic>=2.7.0
numpy>=1.24.3
pyarrow>=14.0.1
typing-inspect==0.9.0
typing_extensions>=4.12.0
together>=0.2.11
langsmith>=0.1.0
langchain-huggingface>=0.0.6
tokenizers>=0.19.1
transformers>=4.39.0
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.2.1+cu118
sentence-transformers>=2.6.0
